Replace CoreTest with WinFormsTest.
