+++ Source/Examples/SharpDX.Core/WinFormsTest/ImGuiNode.cs
