+++ Source/Examples/SharpDX.Core/WinFormsTest/WinFormsTest.csproj
@@ -1,0 +1,38 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+
+  <PropertyGroup>
+    <OutputType>WinExe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <UseWindowsForms>true</UseWindowsForms>
+    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
+    <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\HelixToolkit.SharpDX.Core.Assimp\HelixToolkit.SharpDX.Core.Assimp.csproj" />
+    <ProjectReference Include="..\..\..\HelixToolkit.SharpDX.Core\HelixToolkit.SharpDX.Core.csproj" />
+    <ProjectReference Include="..\..\..\HelixToolkit\HelixToolkit.csproj" />
+  </ItemGroup>
+    <ItemGroup>
+        <Content Include="..\..\..\..\Images\Textures\TextureCheckerboard2.jpg">
+            <Link>TextureCheckerboard2.jpg</Link>
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </Content>
+        <Content Include="..\..\..\..\Images\Textures\TextureCheckerboard2_dot3.jpg">
+            <Link>TextureCheckerboard2_dot3.jpg</Link>
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </Content>
+        <Content Include="..\..\..\..\Images\EnvironmentMaps\Cubemap_Grandcanyon.dds">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </Content>
+        <Content Include="cimgui.dll">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </Content>
+    </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="SharpDX.Desktop">
+            <Version>4.2.0</Version>
+        </PackageReference>
+      <PackageReference Include="ImGui.NET" Version="1.87.3" />
+    </ItemGroup>
+</Project>
